Add torrents counts for Tags and Categories view
